requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cv==1.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
propcache==0.2.0
python-dotenv==1.0.1
pyzbar==0.1.9
qrcode==8.0
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
vt-py==0.18.4
Werkzeug==3.0.4
yarl==1.15.5
zxing==1.0.3
